+++ covsirphy/cleaning/country_data.py
@@ -81,8 +81,8 @@ class CountryData(CleaningBase):
                     - Recovered (int): the number of recovered cases
         """
         if not self.var_dict:
-            s = "Please execute CountryData.set_variables() in advance."
-            raise Exception(s)
+            raise ValueError(
+                "Please execute CountryData.set_variables() in advance.")
         df = self._raw.copy()
         # Rename the columns
         df = df.rename(self.var_dict, axis=1)
