+++ covsirphy/ode/sird.py
@@ -53,6 +53,7 @@ class SIRD(ModelBase):
         self.kappa = kappa
         self.rho = rho
         self.sigma = sigma
+        self.non_param_dict = {"kappa": kappa, "rho": rho, "sigma": sigma}
 
     def __call__(self, t, X):
         """
